Split the model classes out to their own modules. Sorta.
