+++ models/option.py
@@ -1,0 +1,17 @@
+
+def generate_option_class(db):
+    class Option(db.Model):
+        key = db.Column(db.String(100), primary_key=True)
+        value = db.Column(db.String(100), nullable=True)
+
+        def __init__(self, key, value):
+            self.key = key
+            self.value = value
+
+        def to_dict(self):
+            return {
+                'key': self.key,
+                'value': self.value
+            }
+
+    return Option
