+++ run_tests_with_coverage.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
-coverage run --source=tudor,conversions,logic_layer,data_source ./run_tests.py
+coverage run --source=tudor,conversions,logic_layer,data_source,models ./run_tests.py
 coverage html
